Store xtas results as child documents on the parent document rather than as a property (issue #17)
The document type for children is a concatenation of type__task1__task2.
This is not very elegant and it will break if we have more complex pipes
or use parameters. Maybe a hash on the actual pipeline/header is a better
idea?

The child needs to have a proper mapping to specify parent, which is
currently checked before store_single.

I removed the get_all_results functions since it is never used and it caused troubles
if called before the first store_single