No need to fix nondeterminism in nlner_conll
Trying to fix a random seed would still make us depend on implementation
details in nlner_conll. And we're testing xtas here, not nlner, so all
we care about is that the call succeeded and that whichever result nlner
produced was returned correctly. The existing test covers that just fine.